CHG: renamed infrastructure folder and updated persistence entities, mapper, and repository
