+++ .gitignore
@@ -26,7 +26,6 @@ replay_pid*
 # IntelliJ
 .idea/
 *.iml
-out/
 *.iws
 *.ipr
