+++ code.verse/src/main/java/app/code/verse/infrastructure/persistence/entities/EmployeeEntity.java
@@ -1,4 +1,4 @@
-package app.code.verse.infraestructure.persistence.entities;
+package app.code.verse.infrastructure.persistence.entities;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
